backend/requirements.txt
# Core Libraries
numpy==1.26.4
pandas==2.2.1
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.19.0

# Machine Learning
scikit-learn==1.4.1.post1


# Jupyter and Interactive Tools
jupyterlab==4.1.5
ipython==8.22.2
notebook==7.1.2


# Environment Management
python-dotenv==1.0.1
ipykernel==6.29.3


fastapi
uvicorn


openai
requests==2.28.1
